- (touching an empty file so that I can add the notes that I forgot to   add)
Here are the changes that were merged in with rejigger-rx-data

Changes:

- Refill Request Changes : I learned that the requestedPrescription within
a refillRequest is actually the prescription for the refill that the
patient is requesting, and not the prescription from which the refill
request originates. In fact, DoseSpot makes no attempt to link the
refill request to the original prescription. Consequently:

	- We store the requestedPrescription in the requested_treatment table, and
the dispensedPrescription in the pharmacy_dispensed_treatment.

	- We store the doctor_id for the requested_treatment and the
pharmacy_dispensed_treatment because there is a possibility of the
doctor in the pharmacy_dispensed_treatment being different than the
doctor in the requested_treatment.

	- We attempt to link the refill request to the originating prescription on
our end in the following manner:

		a. We try to look for any prescriptions within a 24 hour period (+-12 hour)
around the erx_sent_date received from dosespot, which corresponds to
the time at which the prescription was last fulfillfed.
		b. From within the treatments found in the previous step, we look for a
treatment that matches all the ids to identify the originating
prescription.
		c. If such a prescription is found, then we store originating_treatment_id
in the requested_treatment.

- Treatment Object : Now querying for the doctor, rx_history and pharmacy
associated with a treatment at the same time as getting a treatment
based on id, given that any treatment is prescribed by a particular
doctor and potentially has an rx_history associated with it. This makes
it possible for the client to display all the data as one unit.

- Patient object : Querying for the patient's preferred pharmacy at the
same time as querying for the patient. Again, similar wins to the
treatment object above where it only makes sense to get the patient's
preferred pharmacy as part of the patient object.

- Now storing NDC for drugs : Was previously only storing the drug ids
associated with the particular drug database (lexicomp) but realized its
best that we store the one remaining id that helps uniquely identify the
drug independent of the database :).

- Pharmacy lat/lng as float : It's easier for the client to deal with
lat/lng as float. In fact, its best for us to store the lat/lng as float
as well. Currently storing as string, but have it as a task to store it
as a decimal type so that precision of co-ordinates is maintained in
mysql (Recommended type is decimal(10,6))

Known TODOs

(all have been added to
https://app.asana.com/0/9550210842010/11543132357192)

- Store lat/lng for pharmacy as decimal type

- Currently there is a drug_db_id table for requested_treatment,
pharmacy_dispensed_treatment, and treatment which is sub-optimal.
Rather, it would be better to have 1 drug_db_id table which each of the
different treatment types reference.