Always use TLS even for local dev.
Use predefined cert and key when Debug is set but TLSKey and TLSCert are
not defined. Since in production we'll always be behind TLS (https) it's
nice to also have that guarantee locally. This means that things like
Secure cookies can be used without having to check how the request is
being made.