Unify phone number tables.
The doctor and patient had separate phone numbers tables which were
identical except for the patient one having a status column. There is
now only one phone number table which is linked to the account. This
makes it simpler to lookup a phone number given only the account, and it
allows for other roles such as admin to also have phone numbers.