Better logging.
The purely JSON formatting of the logs before was nice for the log
search indexing, but was terrible for everything else. The default now
is relatively standard syslog style key=value formatting. This makes
the logs much nicer to read.

Another change is to make logging always a string but with optional
context (key value pairs). Setting a context returns a new logger which
allows for passing around context. For instances, requestID could be
attached to every log event if we passed around loggers attached to HTTP
request objects (through http context perhaps).

This change will mess up logsearch for a bit, but I'll sort that out
once this is live.

Another issue solved is that CloudWatch logging wasn't able to parse the
json formatted logs very well so it wasn't possible to make sensible
alerts. With the new format that shouldn't be a problem since it's more
standard.