Enabling the stripe library to function on its own without any dependencies.
Did this by removing the dependency of the stripe library on the payment
package and the common package. The stripe package now has the
definition of all its structs. The restapi no longer works with a
generic PaymentAPI but with a StripeClient which is still an interface.
The reason for keeping the StripeClient as an interface is so that the
tests can easily provide a stubbed out version of the stripe client.

While its preferrable to have a generic way of dealing with a payment
service, the current behavior of the paymentAPI is very strongly linked
to how stripe structures its APIs and data objects. Rather than worrying
about providing a generic payment API for now, decided its better to
consolidate and simplify the stripe package and then introduce a generic
payment API if and when the need arises