awslogidx: change check to be service level rather than agent
Previously awslogidx was creating the check at the agent level.
Unbeknown to me that means that if the check fails the entire host is
considered to have failed. This is obviously bad. This change switches
to creating a service and attaching a check to that instead. This allows
the check to only cause the service (and thus the session) to fail.