Add graduate and undergraduate sections to care provider profile
The MA needs different fields than the doctor, and for now it's easiest
to just add fields instead of making the education more flexible. Making
it flexible either requires a separate table or serialized JSON or other
format in the database. I think it's unlikely we'll need much
flexibility beyond a few fields.