Add 2FA to website login
Includes metrics for auth success and failures.

Two factor on the web works pretty much exactly the same as for the
doctor app. For "device ID" a cookie is stored so that log in from the
same browser doesn't require 2FA more than once. At the moment it's more
or less permanent (as long as the cookie is around) so at some point
we'll want to make it a max of some number of days before requiring 2FA
again.