Split up javascript into modules
This allows for sharing code between apps (e.g. admin and dr
onboarding), and it makes the project much more manageable.

The downside is that a build step is required during development.
resources/build.sh needs to be run each time changes are made to be able
to test them. For then actually deployment build this happens in
travis-ci. It's possible to use a watcher that triggers the build when
files change during development, but I didn't bother with that for now.