Fix security hole in password reset
Don't give feedback on if an email is invalid. Instead, pretend
everything worked as expected but give the person an easy way out
(resend the email or contact support). This way, someone can't use the
endpoint to figure out if certain email addresses belong to patients
(and thus leaking information).