Followup
This PR contains the work directly related to followup. Biggest changes
are the introduction of the SKU to the layout versioning, given that you
can now have multiple layouts for the same condition, with a 1:1 mapping
to a SKU (visit and followup).

Other notable changes are:
- Home card updates to incorporate for followup
- Update to the visit intake layout to move more content to the server
  side
- New API for patient app to get a list of visits
- Update to the doctor's API of getting list of visits to ensure that we
  only return completed visits to doctor app in a patient file
- Introduction of new attachment type (`attachment:visit`) and a title
  field on the attachment to be able to specify text for the attachment
- Test API for the doctor app to trigger followup (note that this will
  be replaced with scheduling of followup which will come later this
week and over the course of next week)

Things yet to come:
-  Scheduling of followup with the ability to specify the message to
   send and the time at which to send the message
- Introduce "free-followup" promotion

Adding SKU to patient visit. Picking up the cost for a visit based on
the SKU versus assuming that we are working with the acne visit

instead of assuming the SKU pick it from the patient visit

remove layout version related tables from the validation so that we don't populate the data in the snapshot

Adding layout support for multiple skus under the same condition

fix test

added support for the new fields in the visit layout structure

WIP: working on followup

Getting followup to work

adding a common way to create a followup visit; adding test for updating layout for followup after it being scheduled

ensuring to return the diagnosis of the previous visit if no diagnosis exists for the followup

publish event when followup visit is created

Supporting AND and OR operations by recursively populating the condition
contents of the condition block

fix test

updating the layout to include the new structures and tags for text to include

adding the context only if the patient has a doctor as part of their care team

Added new layout for followup based on the questions

Inactivating treatment plan for the case when activating a new one

fix tests

generically determining if the visit is a followup visit instead of using the SKU

creating home card states for followup

Inserting a message notification when treatment plan is revised versus
treating it like a new treatment plan has been received by patient

fix test

updating test to capture the expected states of the home cards

Adding support for followup in messaging

reverting back to having just a follow-up type for the link in the message versus a generic visit type

using a bool to indicate whether or not the JSON is templated

fix test

adding text field for message attachments

switching from text to title for mesage attachment

clean up

add comment

moving migrations over to make room for incoming migrations as a result of the rebase

Moving migrations over and validating them again after rebase