Update travis to go 1.4.1 & use older vet
Fixes a number of issues with 1.4:
https://github.com/golang/go/issues?q=milestone%3AGo1.4.1

The latest vet tool includes a change that checks struct tags. The
problem is that the change seems to have a bug that doesn't support
spaces in the tag values. In order to work around this until the tool is
fixed this change checks out an older revision.