Doctor Selection
This PR introduces an API to return a variable list of doctors to the client to pick a doctor from. Rather than relying on heuristics or an algorithm to determine a score for each doctor, we are currently randomly picking doctors based on a pathway,state combination. This is mostly because we don't currently have that many doctors for each pathway,state combination, and partly because doing any scoring now would be over-engineering the solution.

Few things to note:

The patient signup API now supports an optional care_provider_id to specify the doctor picked. The absence of that parameter indicates that the patient chose the fastest available option which also makes the signup call backwards compatible.
The patient visit creation API also supports the care_provider_id to be able to chose your doctor for subsequent cases.
The patient visit and patient signup APIs both support specifying the pathway_id at the time of patient visit creation so that the appropriate layout can be picked for the visit (not something we were doing before). The default assumed here is acne for backwards compatibility reasons.
A change has been made to the routing logic to only pick the doctor part of a case care team to directly route cases to a doctor. Previously we were assuming that if the doctor existed in the team of the patient at the patient file level, then we'd pick that doctor to route cases to. This assumption no longer holds given that we are now giving the patient the ability to select their doctor for every subsequent case they create.
As a result of the update to the routing logic, the promotion logic for routing doctors also had to be updated.
Also added tests to ensure that patient can create a visit (/case) per pathway and the appropriate layout is picked.
The logic for picking doctors to show in the doctor selection API is as follows:

First option will always be the first available option
In the authenticated state (when we know the patient), we check for care teams of previous cases. We identify the doctors in each care team and check whether the doctors are registered to see patients in the incoming (state, pathway_id) combination.
We check if any of these doctors are eligible and pick them first.
If there are still doctors to be picked to meet the desired number of doctors, then we randomly select from the remaining pool of available doctors in the registered states
We determine the imageURLs for the first available by picking from a pool of N doctors where N >>> the required number of images which is now 6