Support resizing media
The new endpoint should likely exist in it's own process to avoid the GC
impact of doing the decoding, resizing, and encoding. It's not a big
deal for now though so leaving it as part of the REST API. Also, one
huge optimization in the future is to use libjpeg-turbo.

- Media URLs now point to an endpoint of the REST API instead of S3
  directly.