Fix image orientation
Fixes https://app.asana.com/0/7876899703933/28057024363225

Previously the image decoder threw out the EXIF data. This change is to
read the EXIF during decode and fix the orientation after resize/crop.
The resulting image has no EXIF data so the color data is slightly
different if orientation is required, but it's not too far off. Fixing
that would probably be quite a bit of work.