analytics: mangle event name
The server validates the event name and is currently throwing out some
events from the client. In order to avoid losing data the server now
mangles the event name to try to make it valid. The reason for
sanitizing the event names is to give some structure to them. They
should be identifiers rather than arbitrary strings. This is to avoid a
wild-west situation with event names.. however, we haven't really sat
down to talk about that so opinions may vary.

The events currently being dropped.
- "install_attribution.Facebook Installs"
- "support.terms_of_use_&_privacy_policy_pressed"