Feedback API
The patient auth endpoint returns an `actions_needed` field which is an
array of objects with a field `type`. For feedback the type id set to
`simple_feedback_prompt`. This tells the app to fetch the feedback
prompt strings. The feedback action is returned when the patient has not
submitted feedback but has an active treatment plan.

The strings returned for the prompt are stored in the `localized_text`
table referenced by tags in `app_text`. This makes it possible to change
them at runtime. A further step could be to add text editing to the
admin.

New API endpoints:

    GET /v1/patient/feedback/simple_prompt

    RESPONSE

    {
        "screen_title": "How did we do?",
        "rating_prompt_text": "Please rate your Spruce experience so far:",
        "comment_placeholder": "Anything else you'd like to tell us?",
        "submit_button_text": "Done"
    }

    POST /v1/patient/feedback

    REQUEST

    {
        "rating": 5,
        "comment": "Best 3v4h!!@!@#!!"
    }

API discussion:
https://github.com/SpruceHealth/api-response-examples/pull/89

Fixes https://app.asana.com/0/search/33568539723433/31559924387163