Fix Consul service registration race
The service registration was not waiting until the registration was
actually completed before returning. This meant that anything that tried
to use the service for sessions (e.g. locks) would fail until the
registration succeeded. While mostly harmless as retries would guarantee
that everythign works eventually, it's not ideal and leads to erroneous
error logging.