Optimize querying of all doctors.
The RX errors worker needs to get a list of all doctors to be able to
query DoseSpot using their clinician ID. The query previously was
expensive as it was pulling the list of IDs and then using the large
multi-join query to get all the doctor info. Since the worker only needs
the basic info (acutally really only the clinician_id but probably not
worth having a DAL function just for that) it's better to just select
all the data at once.