Fix parsing of generic drug names (Latisse)
In general the formatting generic product name that DoseSpot gives us is
in the form "<generic name> <strength> <route> <form>" but that doesn't
hold in some cases like Latisse. This change is to find the earliest
match for strength, route, and form in the formatted name and truncate
it there. This should hopefully work in more cases than currently.