Remove WriteDeveloperError
"DeveloperError" is vague and was used for both client errors (bad
request / validation) and internal errors. Removed it in favor of
explicitly using more appropriate responses. Left
WriteDeveloperErrorWithCode because I'm not sure if the clients are
using the code currently.

There's plenty more that could be done to clean up the error responses,
but for now this should help to clean up the logs.