admin: remove dashboard
The dashboard wasn't being removed and was just adding code to maintain.
Can bring it back if we want it, but likely it would exist in a
different form. While Geckoboard is.. not great.. the one thing I do
like is the idea of pushing data to widgets on a dashboard rather than
the dashboard itself doing any work.