Statically link the restapi binary.
This also changes the behavior of DNS lookups to use the native Go
versions instead of the system libraries. This is the behavior that will
be default on Go 1.5 so it'll be nice to see what the effect is. I
assume it'll be better because the libc dns resolution is a bit crap.

Working with static binaries with Docker is nicer as there's less
dependencies.