Support age restrictions in pathway details
API: https://github.com/SpruceHealth/api-response-examples/pull/96

Both the app and the server use the restrictions. The app checks during
signup if the patient is allowed to start a visit based on the
strictions. It also optionally shows an info modal based on the age. The
server checks the age restrictions during visit creation as a security
measure.

Also:

- moved the "Drugs" page in admin to be under the "Care Coordinator"
  section as it doesn't make sense to be top-level.
- fixed some page layout issues in the admin. "container" classed
  elements shouldn't be nested.