Add home and visit/case statuses for teen visits
When a teen starts a visit they're able to proceed up to but not
including submission. The app will make a call at the end of the intake
to tell the server that the visit is now pending parental consent. At
this point the visit status transitions to PENDING_PARENTAL_CONSENT.
Until a parent grants consent the visit cannot be submitted. Content for
the dialog/modal that the app shows when prompting the teen to message
their parent/guardian for consent comes from the server packaged in the
visit response.

This change does not include the web part of the consent flow. That'll
be addressed in a follow-up PR.

API: https://github.com/SpruceHealth/api-response-examples/pull/97