support publishing error logs to an SNS topic
Can set an SNS topic ARN in the config to which errors of level ERR and
CRIT will be published. Optionally a ratelimiter can be used to throttle
the amount of events that get published.

A different payload is used for email, SNS, and 'default' to try to make
it useable for anyone that subscribes to the topic. The idea is that you
can use email and SNS for alerting, but also a lambda function could
listen and do more advanced functions like post to Slack.