httputil: make logging handler more generic
Remove the dependency on the analytics package from httputil. This makes
the logging handler more useful in general and allows processing of the
request events before logging to add more details (e.g. account ID which
doesn't make sense for httputil to know about, device ID which only
applies to restapi, etc..)

Change request ID to be uint64 rather than int64. It should never be
negative. This doesn't change schemas in the analytics or events DBs,
but that's ok as wrap around is fine and we won't generate an ID
requiring 64-bits for at least a few decades.