Adding ability to check for whether a care provider can treate a patient for a given pathway state combination through the care provider selection api
Summary:
Allowing the client to specify a care_provider_id to the care provider selection API. If the care provider specified is eligible, then the response contains just a single option that represents the particular provider. If provider is not eligible, then a message is specified to indicate to the patient that the doctor is not eligible to treat patient along with regular care provider options.

See: https://phabricator.carefront.net/D3 and https://app.asana.com/0/30298983403801/41106169793539

Reviewers: samuel, janky, michaelraines

Reviewed By: michaelraines

Subscribers: janky

Differential Revision: https://phabricator.carefront.net/D7