Finish contextifiying the restapi
Summary:
This replaces all the apiservice handlers with context aware ones. The big changes are how the old "GetContext" is handled. There's new functions `apiservice.(Must)CtxAccount` and `apiservice.(Must)CtxCache`.

Handlers must wrap themselves with `apiservice.RequestCacheHandler(handler)` in order to get the cache map in their context. The reason this isn't globally applied is because it's not needed everywhere (and requires allocating the map), and because the mutable nature of the cache is scary when thought of globally (and thus should be thought of on a handler level).

Reviewers: kunaljham, michaelraines

Reviewed By: michaelraines

Subscribers: janky

Differential Revision: https://phabricator.carefront.net/D21