Use app events for flagging read messages
Summary:
Was previously doing it on all calls to the message list API.

There's two events, one is for when a specific message is read,
case_message, which is sent by the patient app (though only for the last
message in a case), and the other is all_case_messages which is sent by
the doctor app when messages in a case are read (all messages).

It's a shame to have two differnet events for the patient vs doctor app,
but it's needed for backwards comptibility. The future should be sending
a batch of events (or a list of msg IDs) for each message.

Reviewers: michaelraines, kunaljham

Reviewed By: kunaljham

Subscribers: janky

Differential Revision: https://phabricator.carefront.net/D42