Fix https://app.asana.com/0/12387442313872/46481423999530
Summary:
TreatmentPlanUpdatedEvent was failing fairly frequently in production. It was because the wrong ID was being used to index (parentID instead of contentSourceID) to get the treatment plan, and in certain situations the treatment plan to compare the contents with was not being found.

Fortunately this was only impacting whether or not the treatment plan was considered dirty compared to its content source, and not any patient-doctor interaction.

Reviewers: michaelraines, samuel

Reviewed By: samuel

Subscribers: janky

Differential Revision: https://phabricator.carefront.net/D54