Tokenizing doctor messages
Summary:
Tokens can exist in the scheduled messages or doctor's note to patient. They are either manually entered by the doctor or originate from a previous treatment plan or favorite treatment plan. Given that doctors can edit the tokens, we want to provide doctors with as useful information regarding invalid tokens as possible.

This diff handles token parsing based on the start and end delimiters and attempts to validate tokens present in scheduled messages or a doctor's note.

A tokenized message is validated during treatment plan creation, and the tokens replaced upon submission of treatment plan. The reason we don't replace tokens while the treatment plan is being created is because the doctor might end up favoriting the treatment plan after it is done being created.

Right now the universe of tokens is hard coded in the application layer given there are just 4. Should the pool of possible tokens grow, it would make most sense to make the tokens API more generic, and also separate it out into its own package. The tokens logic can be separated out into its own logic today if we think that would be best.

Reviewers: michaelraines, samuel

Reviewed By: samuel

Subscribers: Hansenq, janky

Differential Revision: https://phabricator.carefront.net/D55