Make it possible to specify resource path
Summary:
Previously we hard coded local path for where to find resources relative to the GOPATH environment variable, based on the project structure of the restapi. Now that we are moving into a world where there will be multiple "projects" (carefinder, curbside, restapi, etc) each with its own location of resources, there is a need to specify the location for where to find resources when looking for them locally.

This change enables you to do that via the `MustInitializeResources` method. The change requires that consumer of resources calls `MustInitializeResources` rather than having it automatically be initialized (as it previously was via the `init()` method). One could argue that this inconvenience of having to call `MustInitializeResources` only to make the local loading of resources is not a good tradeoff, in which case other options I see are to:
- Specify resources path using the `RESOURCEPATH` variable but then its a requirement that `RESOURCEPATH` is set when running restapi locally
- Letting consumer of resources append local resource path to `resource.DefaultBundle` but then you'd have to assume that the `init()`function in the `www` package gets run before the consumer calls it which is fair in the cases we are thinking about, but doesn't feel the cleanest given lack of control.

Reviewers: michaelraines, sibljon, samuel

Reviewed By: samuel

Subscribers: janky

Differential Revision: https://phabricator.carefront.net/D85