Adding a script to update the scheduled messages originating from treatment plans for a given list of treatment plan scheduled messages
Summary:
This script has been added on request by the mbop team to update the scheduled messages in queue originated from treatment plans. There are about 1400 patients that are scheduled to get a message and we want to change the way we are interacting with these patients re: followup. See this: https://docs.google.com/document/d/1voUqzaRoDxU6Oq4mludjMl2cLJ1eo04QU4VUHjCsDX8/edit

I've tested this script locally on a clone of staging. Will test again in staging and force the message to go out to ensure it looks good.

Reviewers: michaelraines, samuel

Reviewed By: samuel

Subscribers: janky

Differential Revision: https://phabricator.carefront.net/D88