Send notification re: message from doctor to email and push
Summary:
This change makes it possible to configure, on a per message basis, notifying patients on multiple channels re: a particular message.
It also requires that EMAIL communication preference is configured for the patient before sending email, so that we have the ability to "unsubscribe" a patient from email notifications if they so email us to do so (currently we don't have the unsubscribe button for email notifications but they can email support for that so wanted to handle it).

The driving force behind this change was to be able to ensure that a patient was notified re: a doctor message and engaged with the app as a consequence.

Reviewers: samuel, michaelraines

Reviewed By: michaelraines

Subscribers: janky

Differential Revision: https://phabricator.carefront.net/D90