Adding ability for anyone with right permissions to update SFTPs for doctors on admin panel
Summary: Given that we have to type our password into PAW and  turn off 2FA to upload, thought it best to add to the admin panel until we have a better way to upload SFTPs. This will also give power to the mbop team to update the SFTPs once they are done updating the content.

Reviewers: sibljon, michaelraines

Reviewed By: michaelraines

Differential Revision: https://phabricator.carefront.net/D109