Updating patient case message attachment types to indicate followup visits vs regular visits
Summary:
We were using different item_types for followup-visit attachments in case messaging, which was causing the email campaign (for 2nd promnpt for followup) to fail as it was not able to find the message based on the item type.

Updating the item_type to be `followup_visit` for all case messages containing a visit as currently we are only attaching a followup visit to case messaging.

Reviewers: michaelraines

Reviewed By: michaelraines

Subscribers: janky

Differential Revision: https://phabricator.carefront.net/D111