Always using the patient_visit as the parent for the treatment plan.
Summary:
Previously we were depending on the client to specify the parent of the treatment plan, which can be either the patient visit or the previous treatment plan. Doing so meant we accepted whatever the client gave us for a piece of information that the server could determine and was not actually driven by user interaction. Secondly, the parent information is only used for metadata purposes and not to drive any logic in the app.

For these reasons, pulling back control of parent determination onto the server, and also using the patient_visit as the parent for now rather than determining the previous treatment plan as that is expensive to do and not really necessary.

In the future, it would probably make most sense to completely remove the parent concept as it is not really needed for the functioning of the product.

Reviewers: sibljon, michaelraines

Reviewed By: michaelraines

Subscribers: janky

Differential Revision: https://phabricator.carefront.net/D120