Allowing drug components to be nullable
Summary: It's possible for the drug name not to be parseable into its components as expected (drug name, route, form). This just happened to a doctor with the medication name: "Derma - Smoothe (Scalp) (topical - oil)". This change makes it so that the drug_form and route are nullable columns in the treatment table.

Reviewers: michaelraines, samuel

Reviewed By: michaelraines, samuel

Subscribers: janky

Differential Revision: https://phabricator.carefront.net/D153