Unify deploy scripts given that the script is the same its just the app name that is different.
Summary:
Realized when trying to put the carefinder binary together that the deploy script was the same for all services, it was just the name that was different. The change here is to have a single deploy script at the cmd/svc level, with the requirement to specify the app name for identifying which service is being deployed.

I know  that we want to soon do away with this deployment process (in favor of the devops service) but this should at least help with avoiding duplication for each service for the short term.

Reviewers: michaelraines, samuel, sibljon

Reviewed By: samuel

Subscribers: janky

Differential Revision: https://phabricator.carefront.net/D199