Bootstrapping CareFinder project + templates for doctor and city pages
Summary:
Purpose of this review is to put the pieces together for the carefinder project. Any request to sprucehealth.com/dermatologist-near-me will be reverse proxied to the carefinder service which will be deployed on the same box as the restapi. Whether or not to reverse proxy requests to the path is controlled by a config variable.

I've also built out the templates + CSS for the city and doctor pages. These can be reached at /dermatologist-near-me/san-francisco-ca.

The next step will be to integrate the database and serve the data that we have been gathering thus far for carefinder.

An overall review of the following items would be much appreciated:
- HTML and CSS for city and doctor pages
- RestAPI service communication with carefinder

Reviewers: sibljon, michaelraines, samuel

Reviewed By: samuel

Subscribers: janky

Differential Revision: https://phabricator.carefront.net/D202