Getting the carefinder data setup for website consumption.
Summary:
Including migration files for the carefinder project. Note that to get the data setup the only way is to restore from snapshot as almost all of the data has been manually inputted into the database.

- migration-1.sql represents a schema dump before the manual data was added
- migration-2.sql are the tables added to store the final carefinder data optimized for quick access of the relevant pieces.

The idea is to work off of a city and (local|spruce) doctor shortlist to identify a city page that exists or a doctor page that exists for the project. Any city or doctor can be made "live" by adding to this list.

Reviewers: michaelraines, samuel

Reviewed By: samuel

Subscribers: janky

Differential Revision: https://phabricator.carefront.net/D225