CareFinder attribution for the CTA.
Summary:
If a user lands upon a spruce doctor profile page via carefinder, the provider_id along with an `is_spruce_patient` flag will be added to branch to route the visit to the doctor as a spruce patient.
If the user clicks on download app from mobile, the branch link will also contain these two flags.

On the Spruce application side, this solution working hinges on the attribution data being hit so that all the attribution parameters are stored for the patient. In this way we will do regular eligibility check if `is_spruce_patient` is provided even if a provider_id is specified, and we will route patient to doctor without practice extension based on this flag.

Reviewers: michaelraines, samuel

Reviewed By: samuel

Subscribers: janky

Differential Revision: https://phabricator.carefront.net/D245