regimensapi: third-party image proxying
Summary:
The Regimens mostly are built using 3rd party images returned from Factual
or scraped from product pages. This means that we're counting on those images
to be available (the links being alive) and of reasonable dimensions. The
only way we can make those guarantees is if we proxy the images. This also
avoids sending our traffic to the 3rd party sites (i.e. hot-linking).
The downside is the extra bandwidth and storage requirements, but this can
be minimized and the upside is worth it. Another benefit of proxying is
avoiding serving non-https content if the viewer is using https.

The solution proposed is to return a custom URL with an ID that maps back
to the actual image URL. The URL points to an endpoint on the regimens API
that returns the original image with optional size paramters for server side
resizing. The first time the endpoint is hit the original image is pulled
and saved and a max-size to save space. If a resize is requested then the
appropriate sized image is generated and cached.

Example:

- Product has image URL http://othersite.com/image.jpg
- The API generates and deterministic ID using a collision resistant hash
  of the URL, checks for a stored status on that ID to detect back URLs which
  can be filtered out, and then stores a mapping of ID -> URL somewhere
  persistent (e.g. DynamoDB) and in cache.
- The API returns a new URL https://api.regimens.co/media/proxy/<ID> as the image URL
- A request by the browser to https://api.regimens.co/media/proxy/<ID> is satisfied as:
  - Check cache for resized image (if requested) and return if found
  - Check store for original image and if found resize and cache if necessary
    and return the original or resized image
  - Pull ID -> URL mapping from cache / persistent-store
    - If not found return placeholder
  - Check stored status of mapping to handle dead links and return placeholder
  - Fetch remote image
    - If fetch fails, record failure status for the ID and return placeholder
    - If fetch succeeds then resize to max stored sized, store persistently
      and resize image and cache if necessary and return image.

Options considered but deemed less desirable:

- Pull images from remote sites and store before returning any URLs to client.
  - CON: This is a non-starter. It's too slow to do at query time, and there's
    too many products in Factual to pre-process.
- Similar to the proposed method, could have an endpoint that takes a URL to
  an image and proxies whatever is asked. The problem is the wide open nature
  which could be abused. This could be solved with signatures, but that adds
  some complexity. Also, it's nicer to have a simple ID than a URL for storing
  cached images.

Reviewers: kunaljham, michaelraines

Reviewed By: michaelraines

Subscribers: sibljon, janky

Differential Revision: https://phabricator.carefront.net/D271