regimens/mediaproxy: speed up URL lookups
Summary:
validate.RemoteHost was doing a DNS lookup for each URL which can be
slow when doing a bulk lookup. Remove it in the fast pass and do that
part of the validation only before fetching as it's mostly a security
check.

Fixed the s3 storage so GetReader properly returns a NotFound error.

Reviewers: kunaljham, michaelraines

Reviewed By: michaelraines

Subscribers: janky

Differential Revision: https://phabricator.carefront.net/D274