Password protect non prod pages
Summary:
Setting up global password protection on the non-prod pages. The admin panel gets wrapped into this as well, which while unfortunate, has the advantage of setting up the filter at a fundamental level without having to set it up for each individual path.

We're planning to move the admin panel to live behind VPN anyway, so this is hopefully temporary where you have to type in a simple password to access the page until we move the admin panel to its own subdomain.

Reviewers: michaelraines, samuel

Reviewed By: samuel

Subscribers: janky

Differential Revision: https://phabricator.carefront.net/D280