Ensuring that automated messages from provider to patient, or between providers via case assignment does not modify the queue of the sender in any way.
Summary:
There is currently a bug in production where an **automated** message from a provider to a patient or another provider (case assignment) modifies the inbox of the sender thinking that the sender explicitly reacted to a pending event (such as a message from a patient or an assignment from another provider).

This fixes that issue by ensuring that automated messages do not impact the queue of the sender at all.

The automated flag could also help in the future if we want to explicitly call out automated messages when they are sent.

This also fixes a bug where the automated message from the sender resulted in an item in the history of the sender, which was weird especially in the cases where a doctor sending a message to the patient resulted in an automated case assignment to the care coordinator.

Reviewers: samuel, michaelraines

Reviewed By: michaelraines

Differential Revision: https://phabricator.corp.sprucehealth.com/D305