Adding service and model definitions for routing and excomms service.
Summary:
- Vendoring grpc/grpc-go, net/http2, golang/protobuf, gogo/protobuf  packages
- New files to review are `svc/routing/svc.proto` and `svc/excomms/svc.proto`

Note: Found it a bit hard to share proto files across directories as the assumption is that they are in the same package. Setup I was going for was `svc/common.proto` for common definitions and then `svc/routing/svc.proto` for routing service specific definitions. I'm sure we could write a script to add the `package X` directive to each common file and then move the generated files into the directory of the service for which we are generating the models, but not sure if worth it yet. See https://github.com/golang/protobuf/issues/39 for more details.

Reviewers: samuel, michaelraines

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D319